requirements.txt
langchain_core==1.1.1
langchain_google_genai==3.2.0
python-dotenv==1.2.1
streamlit==1.52.0
